requirements/pip.txt
boto3==1.28.29
botocore==1.31.29
celery==5.2.7
channels==4.0.0
daphne==4.0.0
dj-database-url==0.5.0
Django==4.2.7
django-cors-headers==4.2.0
django-crontab==0.7.1
django-decouple==2.1
django-filter==23.2
django-rest-passwordreset==1.3.0
django-storages==1.13.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.7
google-api-core==2.11.1
google-api-python-client==2.86.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.60.0
numpy==1.25.2
pandas==2.0.3
Pillow==9.4.0
psycopg2-binary==2.9.7
redis==5.0.0
requests==2.28.1
xhtml2pdf==0.2.9
xmltodict==0.13.0
openpyxl==3.1.2
